split matchfinder.rs from lempziv.rs
