+++ src/orz/lempziv.rs
@@ -11,93 +11,21 @@ pub struct Params {
     pub match_depth_lazy_evaluation2: usize,
 }
 
-#[derive(Copy, Clone)]
-#[repr(packed)]
-pub struct MatchItem {
-    raw1: u16,
-    raw2: u8,
-}
-
 pub struct Encoder {
-    buckets: Vec<EncoderBucket>,
+    buckets: Vec<orz::matchfinder::EncoderMFBucket>,
     mtfs: Vec<orz::mtf::Encoder>,
 }
 
 pub struct Decoder {
-    buckets: Vec<DecoderBucket>,
+    buckets: Vec<orz::matchfinder::DecoderMFBucket>,
     mtfs: Vec<orz::mtf::Decoder>,
-}
-
-struct EncoderBucket {
-    heads: [i16; LZ_BUCKET_ITEM_HASH_SIZE],
-    nexts: [i16; LZ_BUCKET_ITEM_SIZE],
-    items: [u32; LZ_BUCKET_ITEM_SIZE],
-    ring_head: i16,
-}
-
-struct DecoderBucket {
-    items: [u32; LZ_BUCKET_ITEM_SIZE],
-    ring_head: i16,
-}
-
-macro_rules! compute_hash_on_first_4bytes {
-    ($buf:expr, $pos:expr) => {{
-        *$buf.get_unchecked(($pos + 0) as usize) as u32 * 1333337 +
-        *$buf.get_unchecked(($pos + 1) as usize) as u32 * 13337 +
-        *$buf.get_unchecked(($pos + 2) as usize) as u32 * 137 +
-        *$buf.get_unchecked(($pos + 3) as usize) as u32 * 1
-    }}
-}
-
-macro_rules! ring_add {
-    ($a:expr, $b:expr, $ring_size:expr) => {{
-        ($a as usize + $b as usize) % $ring_size as usize
-    }}
-}
-
-macro_rules! ring_sub {
-    ($a:expr, $b:expr, $ring_size:expr) => {{
-        ($a as usize + $ring_size as usize - $b as usize) % $ring_size as usize
-    }}
-}
-
-impl MatchItem {
-    pub fn new_match(index: u16, len: u8) -> MatchItem {
-        MatchItem {
-            raw1: index,
-            raw2: len,
-        }
-    }
-
-    pub fn new_literal(symbol: u8) -> MatchItem {
-        MatchItem {
-            raw1: 0x8000,
-            raw2: symbol,
-        }
-    }
-
-    pub fn get_match_or_literal(&self) -> u8 {
-        (self.raw1 == 0x8000) as u8
-    }
-
-    pub fn get_match_index(&self) -> u16 {
-        self.raw1
-    }
-
-    pub fn get_match_len(&self) -> u8 {
-        self.raw2
-    }
-
-    pub fn get_literal(&self) -> u8 {
-        self.raw2
-    }
 }
 
 impl Encoder {
     pub fn new() -> Encoder {
         Encoder {
             buckets: (0..256)
-                .map(|_| EncoderBucket::new())
+                .map(|_| orz::matchfinder::EncoderMFBucket::new())
                 .collect::<Vec<_>>(),
             mtfs: (0..256)
                 .map(|_| orz::mtf::Encoder::new())
@@ -106,22 +34,17 @@ impl Encoder {
     }
 
     pub fn reset(&mut self) {
-        for bucket in &mut self.buckets {
-            bucket.nexts.iter_mut().for_each(|v| *v = -1);
-            bucket.heads.iter_mut().for_each(|v| *v = -1);
-            bucket.items.iter_mut().for_each(|v| *v = 0);
-            bucket.ring_head = 0;
-        }
+        self.buckets.iter_mut().for_each(|bucket| bucket.reset());
     }
 
     pub unsafe fn encode(&mut self, params: &Params, sbuf: &[u8], tbuf: &mut [u8], spos: usize) -> (usize, usize) {
         let mut spos = spos;
         let mut tpos = 0;
-        let mut match_items = Vec::<MatchItem>::with_capacity(LZ_CHUNK_SIZE);
+        let mut match_items = Vec::<orz::matchfinder::MatchItem>::with_capacity(LZ_CHUNK_SIZE);
 
         // skip first bytes
         if spos == 0 {
-            match_items.push(MatchItem::new_literal(*sbuf.get_unchecked(spos)));
+            match_items.push(orz::matchfinder::MatchItem::new_literal(*sbuf.get_unchecked(spos)));
             spos += 1;
         }
 
@@ -140,7 +63,7 @@ impl Encoder {
                     match_item.get_match_len() as usize,
                     params.match_depth_lazy_evaluation1,
                 ) {
-                    match_item = MatchItem::new_literal(*sbuf.get_unchecked(spos));
+                    match_item = orz::matchfinder::MatchItem::new_literal(*sbuf.get_unchecked(spos));
                 }
             }
             if params.match_depth_lazy_evaluation2 > 0 && match_item.get_match_or_literal() == 0 {
@@ -151,7 +74,7 @@ impl Encoder {
                     match_item.get_match_len() as usize,
                     params.match_depth_lazy_evaluation2,
                 ) {
-                    match_item = MatchItem::new_literal(*sbuf.get_unchecked(spos));
+                    match_item = orz::matchfinder::MatchItem::new_literal(*sbuf.get_unchecked(spos));
                 }
             }
 
@@ -163,7 +86,7 @@ impl Encoder {
                 _ => {
                     let mtf = &mut self.mtfs.get_unchecked_mut(*sbuf.get_unchecked(spos - 1) as usize);
                     let mtf_encoded_literal = mtf.encode(match_item.get_literal());
-                    match_items.push(MatchItem::new_literal(mtf_encoded_literal));
+                    match_items.push(orz::matchfinder::MatchItem::new_literal(mtf_encoded_literal));
                     spos += 1;
                 }
             }
@@ -250,7 +173,7 @@ impl Decoder {
     pub fn new() -> Decoder {
         return Decoder {
             buckets: (0..256)
-                .map(|_| DecoderBucket::new())
+                .map(|_| orz::matchfinder::DecoderMFBucket::new())
                 .collect::<Vec<_>>(),
             mtfs: (0..256)
                 .map(|_| orz::mtf::Decoder::new())
@@ -259,16 +182,13 @@ impl Decoder {
     }
 
     pub fn reset(&mut self) {
-        for bucket in &mut self.buckets {
-            bucket.items.iter_mut().for_each(|v| *v = 0);
-            bucket.ring_head = 0;
-        }
+        self.buckets.iter_mut().for_each(|bucket| bucket.reset());
     }
 
     pub unsafe fn decode(&mut self, tbuf: &[u8], sbuf: &mut [u8], spos: usize) -> Result<(usize, usize), ()> {
         let mut spos = spos;
         let mut tpos = 0;
-        let mut match_items = Vec::<MatchItem>::with_capacity(LZ_CHUNK_SIZE);
+        let mut match_items = Vec::<orz::matchfinder::MatchItem>::with_capacity(LZ_CHUNK_SIZE);
 
         // decode match_items_len
         let match_items_len =
@@ -317,7 +237,7 @@ impl Decoder {
                 }
 
                 if 0 <= b && b < 256 {
-                    MatchItem::new_literal(b as u8)
+                    orz::matchfinder::MatchItem::new_literal(b as u8)
                 } else {
                     let match_index_id = huff_decoder2.decode_from_bits(&mut bits);
                     if match_index_id < 0 || match_index_id >= 32 {
@@ -329,7 +249,10 @@ impl Decoder {
                         match_index_id as usize);
                     let match_index_bits_len = *LZ_MATCH_INDEX_BITS_LEN_ARRAY.get_unchecked_mut(
                         match_index_id as usize);
-                    MatchItem::new_match(match_index_base + bits.get(match_index_bits_len) as u16, match_len)
+
+                    orz::matchfinder::MatchItem::new_match(
+                        match_index_base + bits.get(match_index_bits_len) as u16,
+                        match_len)
                 }
             });
         }
@@ -352,159 +275,27 @@ impl Decoder {
                 _ => {
                     let match_len = match_item.get_match_len() as usize;
                     let match_pos = bucket.get_match_pos(match_item.get_match_index() as i16);
-                    lz_fast_memcopy(sbuf.get_unchecked(match_pos), sbuf.get_unchecked_mut(spos), match_len);
+                    {  // fast increment memcopy
+                        let mut a = sbuf.as_ptr() as usize + match_pos;
+                        let mut b = sbuf.as_ptr() as usize + spos;
+                        let mut l = match_len as isize;
+
+                        while a + 8 > b {
+                            *(b as *mut u64) = *(a as *const u64);
+                            l -= (b - a) as isize;
+                            b += (b - a) as usize;
+                        }
+                        while l > 0 {
+                            *(b as *mut u64) = *(a as *const u64);
+                            l -= 8;
+                            a += 8;
+                            b += 8;
+                        }
+                    }
                     spos += match_len;
                 }
             }
         }
         Ok((spos, tpos))
-    }
-}
-
-impl EncoderBucket {
-    fn new() -> EncoderBucket {
-        EncoderBucket {
-            heads: [-1; LZ_BUCKET_ITEM_HASH_SIZE],
-            nexts: [-1; LZ_BUCKET_ITEM_SIZE],
-            items: [0; LZ_BUCKET_ITEM_SIZE],
-            ring_head: 0,
-        }
-    }
-
-    unsafe fn update_and_match(&mut self, buf: &[u8], pos: usize, match_depth: usize) -> MatchItem {
-        if pos + LZ_MATCH_MAX_LEN + 8 >= buf.len() {
-            return MatchItem::new_literal(*buf.get_unchecked(pos));
-        }
-        let mut max_len = LZ_MATCH_MIN_LEN - 1;
-        let mut max_node = 0;
-
-        let entry = compute_hash_on_first_4bytes!(&buf, pos) as usize % LZ_BUCKET_ITEM_HASH_SIZE;
-        let mut node = *self.heads.get_unchecked(entry);
-
-        // update befault matching (to make it faster)
-        self.ring_head = ring_add!(self.ring_head, 1, LZ_BUCKET_ITEM_SIZE) as i16;
-        *self.nexts.get_unchecked_mut(self.ring_head as usize) = *self.heads.get_unchecked(entry);
-        *self.items.get_unchecked_mut(self.ring_head as usize) = pos as u32 | (*buf.get_unchecked(pos) as u32) << 24;
-        *self.heads.get_unchecked_mut(entry) = self.ring_head as i16;
-
-        // empty node
-        if node == -1 || node == self.ring_head {
-            return MatchItem::new_literal(*buf.get_unchecked(pos));
-        }
-
-        // start matching
-        for _ in 0..match_depth {
-            let (node_first_byte, node_pos) = (
-                (*self.items.get_unchecked(node as usize) >> 24 & 0x000000ff) as u8,
-                (*self.items.get_unchecked(node as usize) >>  0 & 0x00ffffff) as usize);
-
-            if node_first_byte == *buf.get_unchecked(pos) {
-                if *buf.get_unchecked(node_pos + max_len) == *buf.get_unchecked(pos + max_len) {
-                    let lcp = lz_fast_memlcp(
-                        buf.get_unchecked(node_pos) as *const u8,
-                        buf.get_unchecked(pos) as *const u8,
-                        LZ_MATCH_MAX_LEN);
-
-                    if lcp > max_len {
-                        max_node = node;
-                        max_len = lcp;
-                        if max_len == LZ_MATCH_MAX_LEN {
-                            break;
-                        }
-                    }
-                }
-            }
-            node = *self.nexts.get_unchecked(node as usize);
-            if node == -1 || node_pos <= (*self.items.get_unchecked(node as usize) & 0x00ffffff) as usize {
-                break;
-            }
-        }
-
-        if max_len >= LZ_MATCH_MIN_LEN {
-            MatchItem::new_match(
-                ring_sub!(self.ring_head, max_node, LZ_BUCKET_ITEM_SIZE) as u16,
-                max_len as u8,
-            )
-        } else {
-            MatchItem::new_literal(*buf.get_unchecked(pos))
-        }
-    }
-
-    unsafe fn lazy_evaluate(&mut self, buf: &[u8], pos: usize, max_len: usize, depth: usize) -> bool {
-        let entry = compute_hash_on_first_4bytes!(&buf, pos) as usize % LZ_BUCKET_ITEM_HASH_SIZE;
-        let mut node = *self.heads.get_unchecked(entry);
-
-        if node == -1 {
-            return false;
-        }
-        let buf_base = buf.as_ptr() as usize;
-        let buf_cmp_dword = *((buf_base + pos + max_len - 3) as *const u32);
-
-        for _ in 0..depth {
-            let node_pos = (*self.items.get_unchecked(node as usize) >> 0 & 0x00ffffff) as usize;
-            if *((buf_base + node_pos + max_len - 3) as *const u32) == buf_cmp_dword {
-                return true;
-            }
-
-            node = *self.nexts.get_unchecked(node as usize);
-            if node == -1 || node_pos <= (*self.items.get_unchecked(node as usize) & 0x00ffffff) as usize {
-                break;
-            }
-        }
-        return false;
-    }
-}
-
-impl DecoderBucket {
-    fn new() -> DecoderBucket {
-        DecoderBucket {
-            items: [0; LZ_BUCKET_ITEM_SIZE],
-            ring_head: 0,
-        }
-    }
-
-    unsafe fn update(&mut self, pos: usize) {
-        self.ring_head = ring_add!(self.ring_head, 1, LZ_BUCKET_ITEM_SIZE) as i16;
-        *self.items.get_unchecked_mut(self.ring_head as usize) = pos as u32;
-    }
-
-    unsafe fn get_match_pos(&self, match_index: i16) -> usize {
-        let node = ring_sub!(self.ring_head, match_index, LZ_BUCKET_ITEM_SIZE);
-        return *self.items.get_unchecked(node) as usize;
-    }
-}
-
-unsafe fn lz_fast_memlcp(a: *const u8, b: *const u8, max_len: usize) -> usize {
-    let a = a as usize;
-    let b = b as usize;
-    if *(a as *const u32) == *(b as *const u32) {
-        let mut len = 4usize;
-        while len + 4 <= max_len && *((a + len) as *const u32) == *((b + len) as *const u32) {
-            len += 4;
-        }
-
-        // keep max_len=255, so (len + 3 < max_len) is always true
-        len += 2 * (*((a + len) as *const u16) == *((b + len) as *const u16)) as usize;
-        len += 1 * (*((a + len) as *const  u8) == *((b + len) as *const  u8)) as usize;
-        return len;
-    }
-    return 0;
-}
-
-unsafe fn lz_fast_memcopy(a: *const u8, b: *mut u8, len: usize) {
-    let mut a = a as usize;
-    let mut b = b as usize;
-    let mut l = len as isize;
-
-    while a + 8 > b {
-        *(b as *mut u64) = *(a as *const u64);
-        l -= (b - a) as isize;
-        b += (b - a) as usize;
-    }
-    while l > 0 {
-        *(b as *mut u64) = *(a as *const u64);
-        l -= 8;
-        a += 8;
-        b += 8;
     }
 }
